+++ src/entities/resolution.rs
@@ -1,10 +1,24 @@
 use std::str::FromStr;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum UpstreamStrategy {
     RoundRobin,
     Random,
     Failover,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum UpstreamProtocol {
+    Dns,
+    Dot,
+    Doh,
+    Doq,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ResolvedUpstream {
+    pub address: String,
+    pub protocol: UpstreamProtocol,
 }
 
 impl FromStr for UpstreamStrategy {